requirements.txt
anthropic>=0.25.0
playwright>=1.44.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
